Move onetime command code to respective files (#2960)
* rename whitesource command: move to own file

* fix_loop_duplicates command: move logic to own file

* fix imports

* fix imports

* fix_loop_duplicates command: move logic to own file
